Add LOGGER_OPT_PRINTLOST to print the number of lost lines so far ...
Together with the option LOGGER_OPT_NONBLOCK, this option will let you know
how much lines were lost if this happens.

Soon as the buffer is freeed, a line is silently inserted in the flow
telling how much lines were lost so far so that the appropriate action can
be taken.