First implementation of dynamic queues allocation.
A lock is needed to avoid writers to allocate a new queue at the same time,
but this should happen rarely.  The good thing is that each writers are now
able to choose the size of their queue, depending on their real needs.

The only limitation is the number of queues that can be callocated at a time
(threads peaks to foreseen at init).

The new thread ask for a free queue and the queue is allocated / reused in a
smart manner (best fit).  However, if the queue size requested is bigger
than what is already there, a new one will be allocated.

This is still under development and need a lot more testings ...