Fix a 'dead-lock' like race condition in some edge cases ...
If a thread is terminating right after it printed a new line and that last
was added to the write queue in between the check & the sleep of the reader,
then both are waiting for each other infinitely.

This is fixed by checking the reading state in the waiting loop.