Add an option to force the Kernel to really allocate the needed pages
This option slows a bit the allocation of the write queues but speed up the
time spent in logger_printf() when filling the queues for the 1st time.

Without this option, the thread takes 3x to 5x more time to print something
until the queue is completely allocated by the Kernel. So when the queue is
almost completely full the 1st time.

So, if you plan to use big queues, better to use this option to don't loose
time later ...