Add 'LOGGER_OPT_NOQUEUE' option ...
If added to the set of options, when a new thread is swpaned, the queue is
not allocated to the thread.  It will be allocated at the time of the 1st
logger_printf().

This is useful in case you don't know if the thread is going to print
something...  Or of there are really poor chances it prints something ...
Or if it only print an error and then exit.

In such caces, no need to reserve a queue all the time for that ...

This can be set as the default behaviour or per thread.